// src/app/platform-service-api/model/synchronizedEveAccount.ts
/**
 * EveKit Frontend Server
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 2.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { EveKitUserAccount } from './eveKitUserAccount';


/**
 * EveKit synchronized account
 */
export interface SynchronizedEveAccount {
    accessToken?: string;
    refreshToken?: string;
    /**
     * Unique account ID
     */
    aid?: number;
    userAccount?: EveKitUserAccount;
    /**
     * Date (milliseconds UTC) when this account was created
     */
    created?: number;
    /**
     * Account name
     */
    name?: string;
    /**
     * True if this is a character account, false for a corporation account
     */
    characterType?: boolean;
    /**
     * Character ID to use for accessing the EVE XML API
     */
    eveCharacterID?: number;
    /**
     * Character name of character used for access
     */
    eveCharacterName?: string;
    /**
     * Corporation ID of character used for access
     */
    eveCorporationID?: number;
    /**
     * Corporation Name of character used for access
     */
    eveCorporationName?: string;
    /**
     * Date (milliseconds UTC) when this account was last synchronized
     */
    lastSynchronized?: number;
    /**
     * If greater than 0, then the date (milliseconds UTC) when this account was marked for deletion
     */
    markedForDelete?: number;
    /**
     * EVE XML API access key
     */
    eveKey?: number;
    /**
     * EVE XML API access vcode
     */
    eveVCode?: string;
    scopes?: string;
    accessTokenExpiry?: number;
    /**
     * Valid
     */
    valid?: boolean;
}